+++ internal/logger/doc.go
@@ -1,0 +1,2 @@
+// Package logger
+package logger
