+++ utility/alibaba/oss.go
@@ -13,23 +13,23 @@ import (
 // Upload file upload oss
 func Upload(ctx context.Context, fileName, basePath string) error {
 	var (
-		logger   = utility.Helper().Logger(ctx)
-		env, err = env.NewAlibabaEnv(ctx)
+		logger          = utility.Helper().Logger(ctx)
+		alibabaEnv, err = env.NewAlibabaEnv(ctx)
 	)
-	g.Log(logger).Debug(ctx, "env: ", env.String())
+	g.Log(logger).Debug(ctx, "alibabaEnv: ", alibabaEnv.String())
 	if err != nil {
-		g.Log(logger).Error(ctx, "env.NewAlibabaEnv error: ", err)
+		g.Log(logger).Error(ctx, "alibabaEnv.NewAlibabaEnv error: ", err)
 		return err
 	}
 
 	g.Log(logger).Info(ctx, "Upload file to oss fileName:"+fileName+" basePath:"+basePath)
 	// 创建OSSClient实例。
-	client, err := oss.New(env.Endpoint(ctx), env.AccessKeyID(ctx), env.AccessKeySecret(ctx))
+	client, err := oss.New(alibabaEnv.Endpoint(ctx), alibabaEnv.AccessKeyID(ctx), alibabaEnv.AccessKeySecret(ctx))
 	if err != nil {
 		return err
 	}
 	// 获取存储空间。
-	bucket, err := client.Bucket(env.BucketName(ctx))
+	bucket, err := client.Bucket(alibabaEnv.BucketName(ctx))
 	if err != nil {
 		return err
 	}
