+++ utility/cache/doc.go
@@ -1,0 +1,2 @@
+// Package cache is a utility package for cache.
+package cache
