+++ utility/captcha/doc.go
@@ -1,0 +1,2 @@
+// Package captcha is a utility package for captcha.
+package captcha
