+++ utility/doc.go
@@ -1,0 +1,2 @@
+// Package utility is a utility package for utility.
+package utility
